Rename Canvas namespace and folder to Terminal
