+++ GTFO_VR/Core/UI/Terminal/Pointer/PointerEvent.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using UnityEngine;
+﻿using UnityEngine;
 
-namespace GTFO_VR.Core.UI.Canvas.Pointer
+namespace GTFO_VR.Core.UI.Terminal.Pointer
 {
     public class PointerEvent
     {
