+++ GTFO_VR/Core/VR_Input/VRKeyboard.cs
@@ -1,7 +1,5 @@
-﻿using Assets.scripts.KeyboardDefinition;
-using GTFO_VR.Core.PlayerBehaviours;
-using GTFO_VR.Core.UI.Canvas;
-using GTFO_VR.Core.UI.Canvas.KeyboardDefinition;
+﻿using GTFO_VR.Core.PlayerBehaviours;
+using GTFO_VR.Core.UI.Terminal;
 using GTFO_VR.Events;
 using System;
 using System.Text;
