+++ GTFO_VR/GTFO_VR.csproj
@@ -41,20 +41,20 @@
     <Compile Include="Core\PlayerBehaviours\Bhaptics\BhapticsIntegration.cs" />
     <Compile Include="Core\PlayerBehaviours\Bhaptics\BhapticsUtils.cs" />
     <Compile Include="Core\PlayerBehaviours\MovementVignette.cs" />
-    <Compile Include="Core\UI\canvas\Pointer\CanvasPointer.cs" />
-    <Compile Include="Core\UI\canvas\Pointer\MonoPointerEvent.cs" />
-    <Compile Include="Core\UI\canvas\Pointer\PhysicalButton.cs" />
-    <Compile Include="Core\UI\canvas\Pointer\PointerHistory.cs" />
-    <Compile Include="Core\UI\canvas\Pointer\PointerEvent.cs" />
-    <Compile Include="Core\UI\canvas\TerminalReader.cs" />
-    <Compile Include="Core\UI\canvas\KeyboardDefinition\KeyboardLayout.cs" />
-    <Compile Include="Core\UI\canvas\KeyboardDefinition\LayoutParameters.cs" />
-    <Compile Include="Core\UI\canvas\KeyboardDefinition\KeyboardStyle.cs" />
-    <Compile Include="Core\UI\canvas\KeyboardDefinition\KeyDefinition.cs" />
-    <Compile Include="Core\UI\canvas\KeyboardDefinition\LinearLayout.cs" />
-    <Compile Include="Core\UI\canvas\RoundedCubeBackground.cs" />
-    <Compile Include="Core\UI\canvas\TerminalKeyboardCanvas.cs" />
-    <Compile Include="Core\UI\canvas\TerminalKeyboardInterface.cs" />
+    <Compile Include="Core\UI\Terminal\Pointer\TerminalPointer.cs" />
+    <Compile Include="Core\UI\Terminal\Pointer\MonoPointerEvent.cs" />
+    <Compile Include="Core\UI\Terminal\Pointer\PhysicalButton.cs" />
+    <Compile Include="Core\UI\Terminal\Pointer\PointerHistory.cs" />
+    <Compile Include="Core\UI\Terminal\Pointer\PointerEvent.cs" />
+    <Compile Include="Core\UI\Terminal\TerminalReader.cs" />
+    <Compile Include="Core\UI\Terminal\KeyboardDefinition\KeyboardLayout.cs" />
+    <Compile Include="Core\UI\Terminal\KeyboardDefinition\LayoutParameters.cs" />
+    <Compile Include="Core\UI\Terminal\KeyboardDefinition\KeyboardStyle.cs" />
+    <Compile Include="Core\UI\Terminal\KeyboardDefinition\KeyDefinition.cs" />
+    <Compile Include="Core\UI\Terminal\KeyboardDefinition\LinearLayout.cs" />
+    <Compile Include="Core\UI\Terminal\RoundedCubeBackground.cs" />
+    <Compile Include="Core\UI\Terminal\TerminalKeyboardCanvas.cs" />
+    <Compile Include="Core\UI\Terminal\TerminalKeyboardInterface.cs" />
     <Compile Include="Core\UI\RadialItem.cs" />
     <Compile Include="Core\UI\RadialMenu.cs" />
     <Compile Include="Core\PlayerBehaviours\VRMeleeWeapon.cs" />
